server and db connections cleaned up
