+++ .gitignore
@@ -1,3 +1,6 @@
+.DS_Store
+config/default.json
+
 # Logs
 logs
 *.log
