+++ config/connection.js
@@ -1,0 +1,21 @@
+const mongoose = require('mongoose');
+const config = require('config');
+const db = config.get('mongoURI');
+
+const connectDB = async () => {
+  try {
+    await mongoose.connect(db || 'mongodb://localhost/record-crate', {
+      useNewUrlParser: true,
+      useUnifiedTopology: true,
+      useCreateIndex: true,
+      useFindAndModify: false,
+    });
+    console.log('🗄️  DATABASE CONNECTED...');
+  } catch (err) {
+    console.error(err.message);
+    // exit process with failure
+    process.exit(1);
+  }
+};
+
+module.exports = connectDB;
