+++ server.js
@@ -1,7 +1,6 @@
 const express = require('express');
-const mongoose = require('mongoose');
-const config = require('config');
-const db = config.get('mongoURI');
+const connectDB = require('./config/connection');
+const mongoose = require('mongoose'); // can delete before going to production
 
 const app = express();
 const PORT = process.env.PORT || 3001;
@@ -14,14 +13,9 @@ app.use(express.static('public'));
 app.use(require('./routes'));
 
 // connect database
-mongoose.connect(db || 'mongodb://localhost/record-crate', {
-  useFindAndModify: false,
-  useNewUrlParser: true,
-  useUnifiedTopology: true,
-  useCreateIndex: true,
-});
+connectDB();
 
 // Use this to log mongo queries being executed!
-mongoose.set('debug', true);
+mongoose.set('debug', true); // can delete before going to production
 
-app.listen(PORT, () => console.log(`🌍 Connected on localhost:${PORT}`));
+app.listen(PORT, () => console.log(`🌍 Server started on port ${PORT}`));
